fixed a super small bug
Took a long time. Basically, some style setting caused the legacy check or image to get displayed below the blue text.
Fix for this combo was to remove this style. However, this re-introduced a 1 pixel alignment issue because the style aligned the blue and legacy text combo.
Permanent fix was nesting the entire legacy text inside a span to mimic the wrap the blue text code uses.