checkmark selector update is no longer remote
I am leaving the selector hardcoded because it needs to be accessible on install.
I expect that a checkmark update feature will not be often used since it would be off by default due to how intrusive the checkmark cache is. Thus, there is no need to change line 762 in content.js.
Leaving it in this state means a checkmark selector failure would only affect new users.
If this feature gets requested then I will reconsider.